Move models out of assembler
