+++ Elastic.Documentation.Model/Configuration/Repository.cs
