+++ Elastic.Documentation.Model/ContentSource.cs
