+++ Elastic.Documentation.Model/Elastic.Documentation.Model.csproj
@@ -1,0 +1,15 @@
+﻿<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <TargetFramework>net9.0</TargetFramework>
+    <ImplicitUsings>enable</ImplicitUsings>
+    <Nullable>enable</Nullable>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <PackageReference Include="Microsoft.Extensions.Logging"/>
+    <PackageReference Include="NetEscapades.EnumGenerators"/>
+    <PackageReference Include="Vecc.YamlDotNet.Analyzers.StaticGenerator"/>
+    <PackageReference Include="YamlDotNet"/>
+  </ItemGroup>
+</Project>
