+++ Elastic.Documentation.Model/YamlStaticContext.cs
